refactor: 更新 FeedbackController 和相關 DTO，調整評分邏輯及新增焦點、理解和自信度分數

+++ src/main/java/com/learning/api/dto/ChatMessageRequest.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class ChatMessageRequest {
     private Long bookingId;
-    private Byte role;
+    private Integer role;
     private String message;
 }
